+++ module/CLI/src/Command/ShortUrl/ListShortUrlsCommand.php
@@ -10,6 +10,7 @@ use Shlinkio\Shlink\CLI\Util\ExitCode;
 use Shlinkio\Shlink\CLI\Util\ShlinkTable;
 use Shlinkio\Shlink\Common\Paginator\Paginator;
 use Shlinkio\Shlink\Common\Paginator\Util\PagerfantaUtils;
+use Shlinkio\Shlink\Core\Domain\Entity\Domain;
 use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlsParams;
 use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlWithVisitsSummary;
@@ -63,6 +64,12 @@ class ListShortUrlsCommand extends Command
                 'st',
                 InputOption::VALUE_REQUIRED,
                 'A query used to filter results by searching for it on the longUrl and shortCode fields.',
+            )
+            ->addOption(
+                'domain',
+                'd',
+                InputOption::VALUE_REQUIRED,
+                'Used to filter results by domain. Use DEFAULT keyword to filter by default domain',
             )
             ->addOption(
                 'tags',
@@ -134,6 +141,7 @@ class ListShortUrlsCommand extends Command
 
         $page = (int) $input->getOption('page');
         $searchTerm = $input->getOption('search-term');
+        $domain = $input->getOption('domain');
         $tags = $input->getOption('tags');
         $tagsMode = $input->getOption('including-all-tags') === true ? TagsMode::ALL->value : TagsMode::ANY->value;
         $tags = ! empty($tags) ? explode(',', $tags) : [];
@@ -145,6 +153,7 @@ class ListShortUrlsCommand extends Command
 
         $data = [
             ShortUrlsParamsInputFilter::SEARCH_TERM => $searchTerm,
+            ShortUrlsParamsInputFilter::DOMAIN => $domain,
             ShortUrlsParamsInputFilter::TAGS => $tags,
             ShortUrlsParamsInputFilter::TAGS_MODE => $tagsMode,
             ShortUrlsParamsInputFilter::ORDER_BY => $orderBy,
@@ -231,7 +240,7 @@ class ListShortUrlsCommand extends Command
         }
         if ($input->getOption('show-domain')) {
             $columnsMap['Domain'] = static fn (array $_, ShortUrl $shortUrl): string =>
-                $shortUrl->getDomain()?->authority ?? 'DEFAULT';
+                $shortUrl->getDomain()?->authority ?? Domain::DEFAULT_AUTHORITY;
         }
         if ($input->getOption('show-api-key')) {
             $columnsMap['API Key'] = static fn (array $_, ShortUrl $shortUrl): string =>