+++ module/CLI/test-cli/Command/CreateShortUrlTest.php
@@ -8,6 +8,7 @@ use PHPUnit\Framework\Attributes\Test;
 use Shlinkio\Shlink\CLI\Command\ShortUrl\CreateShortUrlCommand;
 use Shlinkio\Shlink\CLI\Command\ShortUrl\ListShortUrlsCommand;
 use Shlinkio\Shlink\CLI\Util\ExitCode;
+use Shlinkio\Shlink\Core\Domain\Entity\Domain;
 use Shlinkio\Shlink\TestUtils\CliTest\CliTestCase;
 
 class CreateShortUrlTest extends CliTestCase
@@ -26,6 +27,6 @@ class CreateShortUrlTest extends CliTestCase
         self::assertStringContainsString('Generated short URL: http://' . $defaultDomain . '/' . $slug, $output);
 
         [$listOutput] = $this->exec([ListShortUrlsCommand::NAME, '--show-domain', '--search-term', $slug]);
-        self::assertStringContainsString('DEFAULT', $listOutput);
+        self::assertStringContainsString(Domain::DEFAULT_AUTHORITY, $listOutput);
     }
 }